refactor: clean up main.go
